+++ resources/gulp/settings.js
@@ -13,7 +13,6 @@ var config = {
 
     paths : {
         node : 'node_modules/',
-        nodeBlender : 'node_modules/blender.js/node_modules/',
         fonts: 'public/fonts/',
         css : {
             resources : 'resources/assets/css/',
@@ -43,7 +42,6 @@ var files = {
         front : {
             sass : 'front/front.scss',
             css : [
-                config.paths.node + 'normalize-css/normalize.css',
                 config.paths.css.resources + 'front.css'
             ],
             js : ['app.js']
@@ -52,10 +50,6 @@ var files = {
         back : {
             sass : 'back/back.scss',
             css : [
-                config.paths.node + 'normalize-css/normalize.css',
-                config.paths.nodeBlender + 'datatables/media/css/jquery.dataTables.css',
-                config.paths.nodeBlender + 'jquery-confirm/css/jquery-confirm.css',
-                config.paths.nodeBlender + '/select2/dist/css/select2.css',
                 config.paths.css.resources + 'back.css'
             ],
             js : ['app.js', 'chart.js']