+++ resources/views/front/home/index.blade.php
@@ -5,12 +5,12 @@
 @section('metaDescription', 'Home view description')
 
 @section('content')
-    <section class="grid">
-        <h1>This is Blender front</h1>
+<main id="home" class="main">
+    <div class="grid">
+        <h1>Blender front</h1>
         <p>
             Lorem ipsum <a href="">dolor sit amet</a>, consectetur adipiscing elit. Phasellus pretium ullamcorper rhoncus. Etiam aliquam efficitur sapien, a aliquam enim lobortis nec. Aliquam eget nisl in sem tincidunt rutrum ut in magna. Pellentesque volutpat dictum arcu at accumsan. Quisque et leo facilisis, condimentum nulla et, porttitor leo. Duis efficitur euismod sapien vel congue. Aliquam nec nisi felis. Donec a aliquam felis, non viverra dui. Nunc ligula arcu, pharetra vitae commodo a, eleifend ac odio. Sed posuere pellentesque eleifend. Aenean bibendum blandit nulla, nec cursus metus tempor at. Etiam a neque arcu. Duis a mi a lectus rutrum accumsan id sed felis. Aenean posuere mi sit amet lobortis gravida. Fusce eu libero lobortis, vulputate diam vitae, aliquet augue. Donec efficitur sem in fermentum molestie.
         </p>
-    </section>
-
-
+    </div>
+</main>
 @endsection